requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
pyjwt
passlib[bcrypt]
python-multipart
